chore(refactor): delete modules folders with 1 file only
