+++ super-agent/src/super_agent/config_storer.rs
@@ -3,4 +3,4 @@ pub mod k8s;
 pub mod loader_storer;
 pub mod store;
 #[cfg(feature = "k8s")]
-pub use k8s::config_map::SubAgentsConfigStoreConfigMap;
+pub use k8s::SubAgentsConfigStoreConfigMap;
