+++ super-agent/tests/k8s/store.rs
@@ -10,7 +10,7 @@ use newrelic_super_agent::k8s::store::{
     K8sStore, StoreKey, CM_NAME_OPAMP_DATA_PREFIX, STORE_KEY_INSTANCE_ID,
     STORE_KEY_OPAMP_DATA_CONFIG_HASH,
 };
-use newrelic_super_agent::opamp::hash_repository::k8s::config_map::HashRepositoryConfigMap;
+use newrelic_super_agent::opamp::hash_repository::k8s::HashRepositoryConfigMap;
 use newrelic_super_agent::opamp::instance_id::{
     getter::{InstanceIDGetter, InstanceIDWithIdentifiersGetter},
     Identifiers,
@@ -25,8 +25,7 @@ use newrelic_super_agent::{
     agent_type::agent_metadata::AgentMetadata, opamp::hash_repository::HashRepository,
 };
 use newrelic_super_agent::{
-    agent_type::agent_values::AgentValues,
-    sub_agent::values::k8s::config_map::ValuesRepositoryConfigMap,
+    agent_type::agent_values::AgentValues, sub_agent::values::k8s::ValuesRepositoryConfigMap,
 };
 use newrelic_super_agent::{
     agent_type::definition::{AgentType, VariableTree},
